\name{launchEditor}
\alias{launchEditor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Launch Hive editor
}
\description{
Call this function and launch a Shiny app with a Hive editor
}
\usage{
launchEditor()
}
%- maybe also 'usage' for other objects documented here.
\author{
Benjamin Chen
}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{
launchEditor()
}
